python/perspective/requirements-dev.txt
backports.shutil-which
codecov>=2.0.15
Faker>=1.0.0
flake8>=3.7.8
ipywidgets>=7.4.2
numpy>=1.8.1
mock
pandas>=0.22.0
psutil>=5.4.8
pyarrow==0.15.0
pytest>=4.3.0
pytest-cov>=2.6.1
pytz>=2018.9
scipy>=1.0.0
six>=1.11.0
Sphinx>=1.8.4
sphinx-markdown-builder>=0.5.2
traitlets>=4.3.2
zerorpc>=0.6.1
